Update MailViewer to Support Inheritance
Inheritance is a key feature of PHP that is very widely used, especially when working with an OOP framework like Laravel. However, Mailviewer currently does not support inheritance when analyzing the constructor of Mailables and ALWAYS throws an error when type-hinting an interface, trait, abstract class, and instantiable classes that may be inherited by child classes. This change adds support for Inheritance to improve useability.